feat(lookup): :sparkles: Add api for lookup list and get owned lists
